+++ NEWS.md
@@ -1,3 +1,7 @@
+# Unreleased version
+
+- Replace `addResourcePath()` with `htmlDependency()` which is more robust
+
 # shinybrowser 1.0.0 (2022-05-11)
 
 Initial CRAN release